Created separate route for profile, created a main component and updated profile choice methods on all distinct page components
Fixes #180
