+++ frontend/course-hub-ui/src/index.js
@@ -9,6 +9,7 @@ import * as serviceWorker from './serviceWorker';
 import 'babel-polyfill';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import CHCompare from "./js/CHCompare";
+import CHProfile from "./js/CHProfile";
 
 ReactDOM.render(
 	<BrowserRouter>
@@ -18,6 +19,7 @@ ReactDOM.render(
 			<Route path="/course" component={CHCourseDetails}/>
 			<Route path="/deals" component={CHDeals}/>
 			<Route path="/compare" component={CHCompare}/>
+			<Route path="/profile" component={CHProfile}/>
 		</Switch>
     </BrowserRouter>, 
 	document.getElementById('root'));
