+++ frontend/course-hub-ui/src/js/CHDeals.js
@@ -178,11 +178,7 @@ class CHDeals extends Component {
 				}
 
 				{choice === "profile" &&
-					[<CHNavigator updateContent={this.handleClick} signedIn={true} caller={"deals"} firstName={firstName} email={email} key="keyNavigatorSearch" />,
-					<div className="profile-content" key="keySearch">
-						<ProfilePage updateContent={this.handleClick} updatePage={this.handlePagination} firstName={firstName} email={email} searchString="" pageNumber={pageNumber}/>
-					</div>,
-					<CHFooter key="keyFooterSearch" />]
+					this.props.history.push('/profile')
 				}
 			</div>
 		);
