+++ frontend/course-hub-ui/src/js/CHSearch.js
@@ -191,14 +191,6 @@ class CHSearch extends Component {
 					<CHFooter key="keyFooterForgotPasswordOverlayOnSearch" />]
 				}
 
-				{choice === "profile" &&
-					[<CHNavigator updateContent={this.handleClick} updatePage={this.handlePagination} signedIn={true} caller={"search"} firstName={firstName} email={email} key="keyNavigatorSearch" />,
-					<div className="profile-content" key="keySearch">
-						<ProfilePage updateContent={this.handleClick} updatePage={this.handlePagination} firstName={firstName} email={email} searchString={searchString} pageNumber={pageNumber}/>
-					</div>,
-					<CHFooter key="keyFooterSearch" />]
-				}
-
 				{choice === "homeSignedIn" &&
 					[<CHNavigator updateContent={this.handleClick} navCallBack={this.toggleModal} updatePage={this.handlePagination} signedIn={firstName != null} caller={"search"} firstName={firstName} email={email} key="keyNavigatorSearch" />,
                      <CHCompareModal updateContent={this.handleClick} isModalOpen={this.state.isOpen} closeModal={this.toggleModal} modalCompareList={this.state.compareList} removeFromModal={this.removeCourseFromModal} />, ,
@@ -210,6 +202,11 @@ class CHSearch extends Component {
 					</div>,
 					<CHFooter key="keyFooterSearch" />]
 				}
+
+				{
+					choice === "profile" && this.props.history.push('/profile')
+				}
+
 				{
 					choice === "coursedetails" && this.props.history.push('/course?courseId=' + this.state.queryString )
 				}
